+++ ee/apps/den-controller/src/auth.ts
@@ -4,7 +4,7 @@ import { emailOTP, organization } from "better-auth/plugins"
 import { APIError } from "better-call"
 import { db } from "./db/index.js"
 import * as schema from "./db/schema.js"
-import { createDenTypeId, normalizeDenTypeId } from "./db/typeid.js"
+import { createDenTypeId, normalizeDenTypeId } from "@openwork-ee/utils/typeid"
 import { sendDenOrganizationInvitationEmail, sendDenVerificationEmail } from "./email.js"
 import { env } from "./env.js"
 import { syncDenSignupContact } from "./loops.js"
