--- ee/apps/den-controller/src/db/typeid.ts
@@ -1,1 +1,0 @@
-export * from "@openwork-ee/utils/typeid"
