+++ ee/apps/den-controller/src/http/desktop-auth.ts
@@ -3,7 +3,7 @@ import express from "express"
 import { and, eq, gt, isNull } from "../db/drizzle.js"
 import { db } from "../db/index.js"
 import { AuthSessionTable, AuthUserTable, DesktopHandoffGrantTable } from "../db/schema.js"
-import { normalizeDenTypeId } from "../db/typeid.js"
+import { normalizeDenTypeId } from "@openwork-ee/utils/typeid"
 import { asyncRoute } from "./errors.js"
 import { getRequestSession } from "./session.js"
 import { z } from "zod"
