+++ ee/packages/den-db/src/index.ts
@@ -1,5 +1,4 @@
-export * from "./client.js"
-export * from "./columns.js"
-export * from "./mysql-config.js"
-export * from "./schema.js"
-export * from "./typeid.js"
+export * from "./client"
+export * from "./columns"
+export * from "./mysql-config"
+export * from "./schema"
