+++ ee/packages/den-db/src/schema/index.ts
@@ -1,0 +1,5 @@
+export * from "./auth"
+export * from "./org"
+export * from "./teams"
+export * from "./workers"
+export * from "./system"
