+++ ee/packages/den-db/src/schema/system.ts
@@ -1,0 +1,26 @@
+import { bigint, int, mysqlTable, uniqueIndex, varchar } from "drizzle-orm/mysql-core"
+import { denTypeIdColumn, timestamps } from "../columns"
+
+export const RateLimitTable = mysqlTable(
+  "rate_limit",
+  {
+    id: denTypeIdColumn("rateLimit", "id").notNull().primaryKey(),
+    key: varchar("key", { length: 512 }).notNull(),
+    count: int("count").notNull().default(0),
+    lastRequest: bigint("last_request", { mode: "number" }).notNull(),
+  },
+  (table) => [uniqueIndex("rate_limit_key").on(table.key)],
+)
+
+export const AdminAllowlistTable = mysqlTable(
+  "admin_allowlist",
+  {
+    id: denTypeIdColumn("adminAllowlist", "id").notNull().primaryKey(),
+    email: varchar("email", { length: 255 }).notNull(),
+    note: varchar("note", { length: 255 }),
+    ...timestamps,
+  },
+  (table) => [uniqueIndex("admin_allowlist_email").on(table.email)],
+)
+
+export const rateLimit = RateLimitTable
