--- ee/packages/den-db/src/typeid.ts
@@ -1,1 +1,0 @@
-export * from "@openwork-ee/utils/typeid"
