+++ ee/packages/utils/src/index.ts
@@ -1,1 +1,1 @@
-export * from "./typeid.js";
+export * from "./typeid"
